# Backend/src/4Wheeler-swagger.yaml
openapi: 3.0.0
info:
  title: 4Wheeler API Documentation
  version: 1.0.0
  description: |-
    API documentation for the 4Wheeler Website.
    This project provides the backend services for managing vehicles for a car dealership, 
    helping customers to have a better look before deciding to go to the store and make a purchase.
    The application offers features such as user authentication, search and sort based on criteria functionality.
    The application is built with scalability, security, and modern REST API design principles in mind.
    Some useful links:
    - [The 4Wheeler_BE repository](https://github.com/Jung2304/4Wheeler_BE)
    - [The source API documentation for the 4Wheeler](https://github.com/Jung2304/4Wheeler_BE/blob/main/Backend/src/4Wheeler-swagger.yaml)
  contact:
    email: 23520333@gm.uit.edu.vn
servers:
  - url: http://localhost:4000
    description: Local server
  - url: https://fourwheeler.onrender.com/api-docs/
    description: Production server
...
paths:
...
  /api/auth/users/refresh-token:
    post:
      tags:
        - auth
      summary: Refreshes the access token
      operationId: refreshToken
      requestBody:
        description: The refresh token using for generating a new access token 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Get new access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Access token refreshed
                  result:
                    $ref: '#/components/schemas/AuthenticateSuccessfully'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Refresh token missing or used"
        '403':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid or expired refresh token"
      security: []
...
  /api/auth/users/forgot-password:
    post:
      tags:
        - auth
      summary: Request password reset via email
      operationId: forgotPassword
      requestBody:
        description: The email to send the OTP code
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: 
                  type: string
                  example: "test@gmail.com"
      responses:
        '200':
          description: OTP sent successfully!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP code sent via email!"
                  OTP:
                    type: string
                    example: "230405"
        '400':
          description: Invalid email format
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation error"
                  errors:
                    type: object
                    properties:
                      email:
                        type: object
                        properties:
                          type:
                            type: string
                            example: "field"
                          value:
                            type: string
                            example: "example@"
                          msg:
                            type: string
                            example: "Invalid or missing email format!"
                          path:
                            type: string
                            example: "email"
                          location:
                            type: string
                            example: "body"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation error"
                  errors:
                    type: object
                    properties:
                      email:
                        type: object
                        properties:
                          type:
                            type: string
                            example: "field"
                          value:
                            type: string
                            example: "Non-existent-email@gmail.com"
                          msg:
                            type: string
                            example: "Email not found!"
                          path:
                            type: string
                            example: "email"
                          location:
                            type: string
                            example: "body"
        '500':
          description: Server error while working with OTP
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error while generating or sending OTP!"
      security: []
...
  /api/auth/users/reset-password:
    post:
      tags:
        - auth
      summary: Reset user's password using reset token
      operationId: resetPassword
      requestBody:
        description: The new password and reset token 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_password: 
                  type: string
                  example: "NewPassword"
                reset_token:
                  type: string
                  example: "7fd6217c4b12644681808f06679d4fd0f4b7dd5fa46f02967ebdd07d733bd991"
      responses:
        '200':
          description: Password reset successfully!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully!"
        '400':
          description: Invalid request or same password as before
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "New password cannot be the same as your old password!"
        '401':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid or expired reset token!"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found!"
        '500':
          description: Server error while resetting password!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error while resetting password!"
      security: []
  /api/auth/users/profile:
    get:
      tags:
        - auth
      summary: Get user profile with favorites
      operationId: getUserProfile
      description: Retrieve the authenticated user's profile information including their favorite cars
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User profile retrieved successfully!"
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "65320a8c32f4e1b2c3d4e5f6"
                      username:
                        type: string
                        example: "testuser1"
                      email:
                        type: string
                        example: "testuser1@gmail.com"
                      phone:
                        type: string
                        example: "0123456789"
                      avatar:
                        type: string
                        example: "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png"
                      role:
                        type: string
                        enum: ["user", "admin"]
                        example: "user"
                      status:
                        type: string
                        example: "active"
                      favorites:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarResponse'
                        description: Array of favorite cars
                      createdAt:
                        type: string
                        format: date-time
                        example: "2023-10-25T12:34:56.789Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2023-10-25T12:34:56.789Z"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found!"
        '500':
          description: Server error while fetching user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error while fetching user profile!"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    AuthenticateSuccessfully:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    CarResponse:
      type: object
      properties:
        _id:
          type: string
          example: "65320a8c32f4e1b2c3d4e5f6"
        make:
          type: string
          example: "BMW"
        model:
          type: string
          example: "X5"
        year:
          type: integer
          example: 2023
        price:
          type: number
          format: double
          example: 75000
        color:
          type: string
          example: "Black"
        category:
          type: string
          example: "SUV"
        seats:
          type: integer
          example: 5
        transmission:
          type: string
          example: "Automatic"
        fuelType:
          type: string
          example: "Gasoline"
        engine:
          type: string
          example: "2.0L Turbo"
        horsepower:
          type: integer
          example: 335
        images:
          type: array
          items:
            type: string
          example: ["image1.jpg", "image2.jpg"]
        description:
          type: string
          example: "Premium luxury SUV with advanced features"
        deleted:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2023-10-25T12:34:56.789Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-25T12:34:56.789Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use the `Bearer` scheme to authenticate requests. Include the token in the `Authorization` header as follows:
        ```
        Authorization: Bearer <your_token>
        ```
security:
  - BearerAuth: []